minipal: move non-critical helpers out of line
Move executable path resolution, uncached thread ID lookup, and thread naming into their source files. Keep only the cached thread ID fast path inline, and configure getauxval availability directly for minipal.
