+++ src/native/minipal/configure.cmake
@@ -12,6 +12,7 @@ check_function_exists(fsync HAVE_FSYNC)
 
 check_symbol_exists(elf_aux_info "sys/auxv.h" HAVE_ELF_AUX_INFO)
 check_symbol_exists(arc4random_buf "stdlib.h" HAVE_ARC4RANDOM_BUF)
+check_symbol_exists(getauxval "sys/auxv.h" HAVE_GETAUXVAL)
 check_symbol_exists(getrandom "sys/random.h" HAVE_GETRANDOM)
 check_symbol_exists(getentropy "unistd.h" HAVE_GETENTROPY)
 check_symbol_exists(O_CLOEXEC fcntl.h HAVE_O_CLOEXEC)
