+++ src/native/minipal/minipalconfig.h.in
@@ -4,6 +4,7 @@
 #cmakedefine01 HAVE_ARC4RANDOM_BUF
 #cmakedefine01 HAVE_GETRANDOM
 #cmakedefine01 HAVE_GETENTROPY
+#cmakedefine01 HAVE_GETAUXVAL
 #cmakedefine01 HAVE_AUXV_HWCAP_H
 #cmakedefine01 HAVE_HWPROBE_H
 #cmakedefine01 HAVE_RESOURCE_H
